Append trailer data to serialized messages.
This reserves space in the Blink envelope for an offset to trailer
data appended to the message which stores summary information about
the message without needing to deserialize it.

This allows quickly assessing which WebIDL interfaces must be exposed
for the message to be deserialized successfully, which means a
messageerror event can be dispatched instead of a message event, even
without eagerly deserializing (as the spec suggests).

This is a major change that requires a bump to the serialization format
and would require care to revert in the future.

Bug: 1341844,1366065
Change-Id: I45904ff3f6270ac0e31c49dfae45c72c7e0d135b