CSS highlight painting: use new spelling/grammar unless feature off
CL:3899651 makes PaintOneSpellingGrammarDecoration take the old code
path (with DocumentMarkerPainter-based squiggly lines) in a couple
of edge cases, in addition to when CSSSpellingGrammarErrors is off.

This patch makes them take the new code path (with the new CSS-based
decorations) in those edge cases, by passing a “decorations override”
to TextDecorationInfo with a static synthesised AppliedTextDecoration
representing the default spelling or grammar decoration, just like how
ComputedStyle::AppliedTextDecorations has a synthesised decoration for
“simple” underlines. We also add crash tests to WPT and wpt_internal
for the SVG case, since CL:3899651 only did so for the other case.

Some callers also needed to be changed, because the assumption that
the decorations in TextDecorationInfo are the same as the decorations
in ComputedStyle is no longer true, but the changes actually make
things simpler.

Like most of the other overrides (but unlike highlight_override_), the
decorations override for a given TextDecorationInfo instance must not
change, because a lot of ctor logic depends on its value (if any).

Bug: 1358810, 1359175
Change-Id: I2b415dec832496dc3a9157dff30d0f6aa44c515f