[last-baseline] Fix inconsistent wm+direction within table internals.
This moves the style adjuster fixup for tables into the layout-tree
builder.

Previously upon the body to root writing mode propagation it was
possible for the style of table internals to be inconsistent. By
moving the fixup into the tree-builder we can ensure it is correct,
and has the side effect of reporting the correct style from computed
style.

Fixed: 1366340, 736093
Change-Id: I261010f8eb6de051a4a8ae112303064c9e426a15