[CompositeClipPathAnimations] Make paint rect encompass entire animation
Previously, the paint rect for a composited clip path animation would be
based on the size of the clip path at paint time, which meant that a
clip path animation progressing from a smaller to larger keyframe would
be clipped unnecessarily. This change decides the clip area based on the
union of all keyframes, resulting in a rect that encompasses the entire
animation.

Fuzzy diff is added for one test because of rendering code path differences causing a very small diff. See discussion on crbug.com/1249071 for why this is preferred.

Bug: 1248610
Change-Id: Ic9f5658166cbe5698b3486fc6f97fc8a4405850c