[FlexNG] RelayoutAndBreakEarlier stack overflow
Similar to CL:3582397, if we RelayoutIgnoringChildScrollbarChanges()
and then RelayoutAndBreakEarlier(), we would hit a stack overflow
because RelayoutAndBreakEarlier() doesn't set
`ignore_child_scrollbar_changes_` (which leads us to alternate between
the two relayouts infinitely).

To fix this, create a new RelayoutAndBreakEarlier() overload in
NGLayoutAlgorithm that takes and Algorithm as input. This allows us
to pass in an Algorithm to RelayoutAndBreakEarlier() from
NGFlexLayoutAlgorithm that sets `ignore_child_scrollbar_changes_`
correctly.

Bug: 1370404
Change-Id: I9a98749d30e55cdc2aad3c8e001e2e226354b509