[GridNG] Subgrid GetComputedStyle and filtering
This change completes basic lined-name support for subgrids. There are
two main parts of this change: updating to serialization of subgrid
named grid lines in accordance with the spec, and a second piece that
filters out named lines from the parent that are out of the subgrid's
range. Line names are not inherited for the non-subgridded axis. There
wasn't an existing test validating this condition, so I added one.

The basic concept is that all named lines are relative to index 0
internally. This is how non-subgrids currently work. During placement,
line names are resolved relative to index 0, and then get shifted by
the subgrid's position.

This also fixes a few various issues - the proper parent grid is
passed in for inheriting line names through nested subgrids. If
the line name filtering results in an empty merged map, the line
name should be removed from the map.

This change regresses `parent-repeat-auto-fit-001.html` and
`parent-repeat-auto-fit-002.html`. However, these were never rendering
correctly in the first place, so I am not sure how they were marked
as [Pass]. I will fix these in a follow-up, by handling
`auto_repeat_named_grid_lines` shared from the parent grid.

Bug: 618969
Change-Id: I1fdcfbffcaedf715a3b0ceb8b69ca709cdee2128